+++ apps/games/lib/hooks.ts
